Moved GenereicDirectoryScreen to components
